% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contactdata-package.R
\docType{package}
\name{contactdata-package}
\alias{contactdata}
\alias{contactdata-package}
\title{contactdata: Social Contact Matrices for 177 Countries}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Data package for the supplementary data in Prem et al. (2017) \doi{10.1371/journal.pcbi.1005697} and Prem et al. \doi{10.1371/journal.pcbi.1009098}. Provides easy access to contact data for 177 countries, for use in epidemiological, demographic or social sciences research.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hugogruson.fr/contactdata/}
  \item \url{https://github.com/bisaloo/contactdata}
  \item Report bugs at \url{https://github.com/bisaloo/contactdata/issues}
}

}
\author{
\strong{Maintainer}: Hugo Gruson \email{hugo.gruson+R@normalesup.org} (\href{https://orcid.org/0000-0002-4094-1476}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Kiesha Prem (\href{https://orcid.org/0000-0003-0528-798X}{ORCID}) [data contributor]
  \item Alex Richard Cook (\href{https://orcid.org/0000-0002-6271-5832}{ORCID}) [data contributor]
  \item Mark Jit (\href{https://orcid.org/0000-0001-6658-8255}{ORCID}) [data contributor]
}

}
\keyword{internal}
