% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_countries.R
\name{list_countries}
\alias{list_countries}
\title{Get the list of countries included in the dataset}
\usage{
list_countries(data_source = c("2020", "2017"))
}
\arguments{
\item{data_source}{Character. Either "202O" (default) or "2017"}
}
\value{
A character vector with the name of all countries included in the
dataset
}
\description{
Get the list of countries included in the dataset
}
\note{
This package uses the nomenclature from the \pkg{countrycode} package.
If your names differ from the names used here, you should use
\pkg{countrycode} as well to update them.
}
\examples{
list_countries()

}
\references{
Kiesha Prem, Alex R. Cook, Mark Jit, Projecting social contact
matrices in 152 countries using contact surveys and demographic data, PLoS
Comp. Biol. (2017), \doi{10.1371/journal.pcbi.1005697}
}
